+++ tests/it/io/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "io_print")]
+mod print;
