+++ tests/it/main.rs
@@ -2,3 +2,5 @@ mod alloc;
 mod bitmap;
 mod buffer;
 mod ffi;
+
+mod io;
